Remove `id` field from `Resource` (#191)
Related to #188 

API resources define a common `id` field, which requires additional
bookkeeping in `_prepare_model`. However, these fields aren't used, and
either shadow a field on the resource or can be computed. This PR
removes `id` from the base model, adds properties for subclasses that
don't have an `id` field (like `Hardware` and `Collection`), and marks
them as deprecated, for anyone who is relying on that behavior.

---------

Signed-off-by: Mattt Zmuda <mattt@replicate.com>